requirements.txt
pandas>=2.0
openpyxl>=3.1
streamlit>=1.32
python-dotenv>=1.0
